requirements.txt
discord.py==2.1.0
beautifulsoup4==4.10.0
bs4==0.0.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus
dnspython==1.16.0
humanfriendly==9.2
motor==2.3.1
python-dotenv